python client library clean branch for PR. #455
