+++ .gitignore
@@ -5,11 +5,10 @@ core
 .DS_Store
 *.swp
 __pycache__
-python/web/current
-python/web/src/rascsi_interface_pb2.py
-python/oled/current
-python/oled/src/rascsi_interface_pb2.py
+current
+rascsi_interface_pb2.py
 src/raspberrypi/hfdisk/
 *~
 messages.pot
 messages.mo
+s.sh
