+++ python/common/requirements.txt
@@ -1,0 +1,2 @@
+protobuf==3.19.3
+requests==2.26.0
